Improvements about users controller
Mainly these improvements are about the users requests/controller
methods.

But, there are small extras too. For instance, the `.env_dist` file was
added as the file having the ENV example vars. This will be used to
contain the ENV vars on the repo without showing their values.
For this to work, you need to create a `.env` file on the project
folder and require the necessary data. For now, I just put there the
PORT number.